# backend/services/merchant-service/api-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "支付平台商户管理服务API文档",
        "title": "Merchant Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@payment-platform.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/security/2fa/disable": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security"
                ],
                "summary": "禁用2FA",
                "parameters": [
                    {
                        "description": "禁用请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Disable2FARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/security/2fa/verify": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security"
                ],
                "summary": "验证2FA",
                "parameters": [
                    {
                        "description": "验证请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Verify2FARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/security/login-activities": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security"
                ],
                "summary": "获取登录活动",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/security/password": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/security/sessions/{session_id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security"
                ],
                "summary": "撤销会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "会话ID",
                        "name": "session_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/security/settings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security"
                ],
                "summary": "获取安全设置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Security"
                ],
                "summary": "更新安全设置",
                "parameters": [
                    {
                        "description": "安全设置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment-platform_merchant-service_internal_service.UpdateSecuritySettingsInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_handler.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "minLength": 8
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
        "internal_handler.Disable2FARequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.Verify2FARequest": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
...
        "payment-platform_merchant-service_internal_service.UpdateSecuritySettingsInput": {
            "type": "object",
            "properties": {
                "abnormal_notification": {
                    "type": "boolean"
                },
                "allowed_countries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "blocked_countries": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ip_whitelist": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "login_notification": {
                    "type": "boolean"
                },
                "max_concurrent_sessions": {
                    "type": "integer"
                },
                "password_expiry_days": {
                    "type": "integer"
                },
                "session_timeout_minutes": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}